requirements.txt
beautifulsoup4
pandas
requests
psycopg2 
python-dotenv
dash
dash-bootstrap-components
plotly